- error.ts moved to a folder - TypeScript types are changed again
