+++ types/index.d.ts
@@ -5,13 +5,33 @@
  * If you need something basic then you should install the Easy Auth application from Wix App Market and use pre-built widgets with your own setup.
  */
 declare module '@exweiv/easy-auth' {
+    namespace Facebook {
+        type RedirectURLOptions = {
+            redirect_uri: string,
+            client_id: string,
+            response_type: string,
+            scope: string,
+            state: string,
+        }
+
+        type AuthOptions = {
+            client_id: string,
+            client_secret?: string,
+            redirect_uri: string,
+            code: string,
+            fields?: string
+        }
+
+        type AuthResponse = { [key: string]: any };
+    }
+
     interface facebook {
         /**
          * Creates a redirect url for authenticating user via Facebook
          * 
          * @param options Options that's used when creating redirect url.
          */
-        redirectURL(options: facebook["FacebookRedirectURLOptions"]): string;
+        redirectURL(options: Facebook.RedirectURLOptions): string;
 
         /**
          * Gets user data from Facebook
@@ -20,24 +40,14 @@ declare module '@exweiv/easy-auth' {
          * @param getClientSecret Defaults to true when set to false you can pass your own client_secret when set to true API will try to fetch client_secret from Wix Secrets Manager, create a secret with `FacebookClientSecret` name.
          * @param access_token Defaults to undefined, when you pass a access_token this access_token will be used when getting user data. Otherwise API will generate new one.
          */
-        userAuth(options: facebook["FacebookAuthOptions"], getClientSecret?: boolean, access_token?: string): Promise<facebook["FacebookAuthResponse"]>;
+        userAuth(options: Facebook.AuthOptions, getClientSecret?: boolean, access_token?: string): Promise<Facebook.AuthResponse>;
+    }
 
-        FacebookRedirectURLOptions: {
-            redirect_uri: string,
-            client_id: string,
-            response_type: string,
-            scope: string,
-            state: string,
-        }
+    namespace Google {
 
-        FacebookAuthOptions: {
-            client_id: string,
-            client_secret?: string,
-            redirect_uri: string,
-            code: string,
-            fields?: string
-        }
+    }
 
-        FacebookAuthResponse: { [key: string]: any };
+    interface google {
+        
     }
 }
